Dramatic refactoring of the cases
Since we removed the ordering constraint required by printing the operation before variable input, we can now dramatically refactor that code to be written once rather than repeated (see: the DRY principle)
This also allows a pretty impressive reduction in the add, mult, etc. functions. In cases where a function simply returns the curly brackets are still required. As long as it's not tremendously long, I'm okay with putting that all on one line.