Extract duplicate code into a function
The operation switch statements had converged, so I extracted them into a shared function which considers the input operation, a, and b. Note: this function is pure except in circumstances where invalid data is allowed to reach it. So long as the input predicates remain in place, this should never occur.

This operation allowed significant cleanup, but did require handling the case where the input was not valid (handled with exit for now, exceptions later?) and moving the `enum Operation` to global scope.