Redistribute variables
Some variables must survive loop iterations, so they're declared in a higher scope. Some can be reset every time, so they can be declared in the lower scope. It's acceptable in C++ (but again, not in C) to declare and initialize a variable on one line, as long as you keep your methods short and readable. Declarations without initialization still belong at the top of the function.